Update 0.1
Pushing nearly finished game.
1d platformer where you shoot through enemies in space to reach the end game.
Plans to add script to game to help it along.

Known issues: Enemies when shot close to turning points are stuck on the barrier because the contact never closed.
Players are not all available for selection as some are missing crucial HUD assets that are currently unfinished. (Only "Dave" is playable at current time)